# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Berlin MarketHUB API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  draft: 'draft',
  active: 'active',
  sold: 'sold',
  archived: 'archived',
  pending_review: 'pending_review',
} as const;

export interface Listing {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  images: string[];
  category: string;
  condition?: string;
  status: ListingStatus;
  tags?: string[];
  /** @nullable */
  shippingFee?: number | null;
  quantity: number;
  sellerId: string;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  sellerAvatar?: string | null;
  sellerVerified: boolean;
  likesCount: number;
  viewsCount: number;
  soldCount: number;
  /** @nullable */
  rating?: number | null;
  isLiked: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  images: string[];
  category: string;
  condition?: string;
  tags?: string[];
  /** @nullable */
  shippingFee?: number | null;
  quantity: number;
  status?: string;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  images?: string[];
  category?: string;
  condition?: string;
  tags?: string[];
  /** @nullable */
  shippingFee?: number | null;
  quantity?: number;
  status?: string;
}

export interface LikeStatus {
  liked: boolean;
  likesCount: number;
}

export interface Category {
  id: string;
  name: string;
  icon: string;
  count: number;
}

export interface Badge {
  id: string;
  name: string;
  icon: string;
  earnedAt: string;
}

export interface Profile {
  id: string;
  username: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  isVerified: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  followersCount: number;
  followingCount: number;
  listingsCount: number;
  /** @nullable */
  totalSales?: number | null;
  /** @nullable */
  responseTime?: string | null;
  joinedAt: string;
  isFollowing: boolean;
  badges: Badge[];
}

export interface ProfileUpdate {
  username?: string;
  fullName?: string;
  bio?: string;
  avatarUrl?: string;
  country?: string;
  location?: string;
  websiteUrl?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
}

export interface FollowStatus {
  following: boolean;
  followersCount: number;
}

export interface SellerStats {
  totalSales: number;
  itemsSold: number;
  successRate: number;
  responseTime: string;
  repeatBuyers: number;
  positiveReviews: number;
  rating: number;
}

export interface Review {
  id: string;
  /** @nullable */
  listingId?: string | null;
  sellerId: string;
  reviewerId: string;
  reviewerName: string;
  /** @nullable */
  reviewerAvatar?: string | null;
  rating: number;
  comment: string;
  isVerifiedPurchase: boolean;
  helpfulCount: number;
  createdAt: string;
}

export interface ReviewInput {
  /** @nullable */
  listingId?: string | null;
  sellerId: string;
  rating: number;
  comment: string;
}

export interface CartItem {
  id: string;
  listingId: string;
  title: string;
  price: number;
  image: string;
  quantity: number;
  sellerId: string;
  sellerName: string;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shippingTotal: number;
  total: number;
}

export interface CartItemInput {
  listingId: string;
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Order {
  id: string;
  listingId: string;
  listingTitle: string;
  listingImage: string;
  price: number;
  buyerId: string;
  sellerId: string;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  sellerName?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export interface HomeDiscovery {
  featured: Listing[];
  trending: Listing[];
  categories: Category[];
  recentlyViewed: Listing[];
}

export interface SearchResults {
  listings: Listing[];
  total: number;
  suggestions: string[];
}

export interface ImageUploadInput {
  dataUrl: string;
  bucket?: string;
  filename?: string;
}

export interface UploadResult {
  url: string;
  path: string;
  size: number;
}

export interface PeriodRevenue {
  period: string;
  revenue: number;
  orders: number;
}

export interface CategoryStats {
  category: string;
  revenue: number;
  count: number;
}

export interface ProductStats {
  listingId: string;
  title: string;
  sold: number;
  revenue: number;
}

export interface SellerAnalytics {
  totalRevenue: number;
  totalOrders: number;
  conversionRate: number;
  avgResponseTime: number;
  refundRate: number;
  repeatBuyerRate: number;
  revenueByPeriod: PeriodRevenue[];
  topCategories: CategoryStats[];
  bestSellingProducts: ProductStats[];
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  message: 'message',
  follow: 'follow',
  like: 'like',
  order: 'order',
  review: 'review',
  price_drop: 'price_drop',
  promo: 'promo',
} as const;

export type NotificationData = { [key: string]: unknown };

export interface Notification {
  id: string;
  type: NotificationType;
  title: string;
  body: string;
  isRead: boolean;
  data?: NotificationData;
  createdAt: string;
}

export interface NotificationReadInput {
  ids?: string[];
  all?: boolean;
}

export type ListListingsParams = {
category?: string;
search?: string;
sellerId?: string;
minPrice?: number;
maxPrice?: number;
condition?: string;
sort?: ListListingsSort;
limit?: number;
offset?: number;
};

export type ListListingsSort = typeof ListListingsSort[keyof typeof ListListingsSort];


export const ListListingsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
} as const;

export type ListFeaturedListingsParams = {
category?: string;
limit?: number;
};

export type ListReviewsParams = {
listingId?: string;
sellerId?: string;
limit?: number;
};

export type ListOrdersParams = {
type?: ListOrdersType;
};

export type ListOrdersType = typeof ListOrdersType[keyof typeof ListOrdersType];


export const ListOrdersType = {
  purchases: 'purchases',
  sales: 'sales',
} as const;

export type SearchListingsParams = {
q?: string;
category?: string;
minPrice?: number;
maxPrice?: number;
sort?: string;
};

export type ListNotificationsParams = {
limit?: number;
unreadOnly?: boolean;
};