# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Berlin MarketHUB API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List all product listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: condition
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/featured:
    get:
      operationId: listFeaturedListings
      tags: [listings]
      summary: Get featured/highlighted listings for home feed
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Featured listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
...
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get a listing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /listings/{id}/like:
    post:
      operationId: toggleListingLike
      tags: [listings]
      summary: Toggle like/wishlist on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatus"
  /listings/{id}/sold:
    patch:
      operationId: markListingAsSold
      tags: [listings]
      summary: Mark a listing as sold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /discovery/search:
    get:
      operationId: searchListings
      tags: [discovery]
      summary: Search listings with filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        category:
          type: string
        condition:
          type: string
        status:
          type: string
          enum: [draft, active, sold, archived, pending_review]
        tags:
          type: array
          items:
            type: string
        shippingFee:
          type: ["number", "null"]
        quantity:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: ["string", "null"]
        sellerAvatar:
          type: ["string", "null"]
        sellerVerified:
          type: boolean
        likesCount:
          type: integer
        viewsCount:
          type: integer
        soldCount:
          type: integer
        rating:
          type: ["number", "null"]
        isLiked:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - price
        - category
        - status
        - sellerId
        - likesCount
        - viewsCount
        - soldCount
        - isLiked
        - images
        - quantity
        - sellerVerified
        - createdAt
        - updatedAt
    ListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        category:
          type: string
        condition:
          type: string
        tags:
          type: array
          items:
            type: string
        shippingFee:
          type: ["number", "null"]
        quantity:
          type: integer
        status:
          type: string
      required:
        - title
        - price
        - category
        - images
        - quantity
    ListingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        category:
          type: string
        condition:
          type: string
        tags:
          type: array
          items:
            type: string
        shippingFee:
          type: ["number", "null"]
        quantity:
          type: integer
        status:
          type: string
    LikeStatus:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required:
        - liked
        - likesCount
...
    SearchResults:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        suggestions:
          type: array
          items:
            type: string
      required:
        - listings
        - total
        - suggestions
...